requirements.txt
# Core dependencies
streamlit>=1.39.0
streamlit-aggrid
crewai>=0.76.9
anthropic>=0.37.1
pydantic>=2.9.2
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
num2words>=0.5.13

# Audio processing
soundfile>=0.12.1
librosa>=0.10.1
numpy>=1.24.0
scipy>=1.11.0

# TTS Models
transformers>=4.35.0
torch==2.3.0
torchaudio==2.3.0
accelerate>=0.26.0

# Development
black>=23.0.0
isort>=5.12.0
pytest>=7.0.0
pytest-cov>=4.1.0
